feat: Generalize test case to run any test method by name
