+++ README.md
@@ -6,7 +6,7 @@ This project is a TDD implementation of the xUnit testing framework using pure J
 
 Supported features:
 
-- [ ] Run test method
+- [x] Run test method
 - [ ] Run setUp() before test method
 - [ ] Run tearDown() after test method
     - [ ] Must be run even if the test method fails
