+++ src/xunit/XUnitTest.java
@@ -3,10 +3,6 @@ package xunit;
 public class XUnitTest {
 
     public static void main(String[] args) {
-        final WasRun wasRun = new WasRun("testMethod");
-
-        System.out.println(wasRun.wasRun()); // false
-        wasRun.run();
-        System.out.println(wasRun.wasRun()); // true
+        new TestCaseTest("testRunning").run();
     }
 }
